solutions/max_kaskevich/6/sources/multicast_communication/quote_message: fixed warning int to double cast.

+++ solutions/max_kaskevich/6/sources/multicast_communication/quote_message.h
@@ -26,7 +26,7 @@ namespace multicast_communication
 
 	public:
 
-        static const std::map< const char, uint32_t > denominator_map;
+        static const std::map< const char, double > denominator_map;
 
         enum quote_type
         {
